requirements.txt
PyQt5==5.15.11
pyserial==3.5
colorama==0.4.6